requirements.txt
pytest~=7.1.1
pytest-ordering
pytest-xdist
pytest-rerunfailures
pytest-instafail
pytest-html
allure-pytest
PyYAML
requests
selenium
xlrd
zmail